+++ BuildConsensus.py
@@ -6,27 +6,25 @@ Builds a consensus sequence for each set of input sequences
 __author__    = 'Jason Anthony Vander Heiden'
 __copyright__ = 'Copyright 2013 Kleinstein Lab, Yale University. All rights reserved.'
 __license__   = 'Creative Commons Attribution-NonCommercial-ShareAlike 3.0 Unported'
-__version__   = '0.4.6'
-__date__      = '2015.05.13'
+__version__   = '0.4.7'
+__date__      = '2015.05.26'
 
 # Imports
 import os, sys, textwrap
 from argparse import ArgumentParser
 from collections import OrderedDict
-from itertools import izip, izip_longest
-from Bio.Seq import Seq
-from Bio.SeqRecord import SeqRecord
-from Bio.Alphabet import IUPAC
+from itertools import izip
 
 # IgPipeline imports
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
-from IgCore import default_delimiter, default_out_args, default_missing_chars
+from IgCore import default_delimiter, default_out_args
 from IgCore import default_barcode_field, default_min_freq
+from IgCore import deleteSeqPositions, findGapPositions, calculateSetError
 from IgCore import annotationConsensus, getAnnotationValues
 from IgCore import flattenAnnotation, mergeAnnotation
 from IgCore import CommonHelpFormatter, getCommonArgParser, parseCommonArgs 
 from IgCore import getFileType, printLog 
-from IgCore import getScoreDict, frequencyConsensus, qualityConsensus
+from IgCore import frequencyConsensus, qualityConsensus
 from IgCore import calculateDiversity, indexSeqSets, subsetSeqSet
 from IgCore import collectSeqQueue, feedSeqQueue
 from IgCore import manageProcesses, SeqResult
@@ -36,93 +34,7 @@ default_min_count = 1
 default_min_qual = 0
 
 
-def deleteSeqPositions(seq, positions):
-    """
-    Deletes a list of positions from a SeqRecord
-
-    Arguments:
-    seq = a SeqRecord objects
-    positions = a set of positions (indices) to delete
-
-    Returns:
-    a modified SeqRecord with the specified positions removed
-    """
-    seq_del = ''.join([x for i, x in enumerate(seq.seq) if i not in positions])
-    record = SeqRecord(Seq(seq_del, IUPAC.ambiguous_dna),
-                       id=seq.id, name=seq.name, description=seq.description)
-
-    if 'phred_quality' in seq.letter_annotations:
-        qual_del = [x for i, x in enumerate(seq.letter_annotations['phred_quality']) \
-                    if i not in positions]
-        record.letter_annotations['phred_quality'] = qual_del
-
-    return record
-
-
-def findGapPositions(seq_list, max_gap, gap_chars=set(['.', '-'])):
-    """
-    Finds positions in a set of aligned sequences with a high number of gap characters.
-
-    Arguments:
-    seq_list = a list of SeqRecord objects with aligned sequences
-    max_gap = a float of the maximum gap frequency to consider a position as non-gapped
-    gap_chars = set of characters to consider as gaps
-
-    Returns:
-    a list of positions (indices) with gap frequency greater than max_gap
-    """
-    # Return an empty list in the singleton case
-    seq_count = float(len(seq_list))
-    if seq_count == 1:
-        return []
-
-    # Iterate through positions and count gaps
-    gap_positions = []
-    seq_str = [str(s.seq) for s in seq_list]
-    for i, chars in enumerate(izip_longest(*seq_str, fillvalue='-')):
-        gap_count = sum([chars.count(c) for c in gap_chars])
-        gap_freq = gap_count / seq_count
-
-        # Update gap position over threshold
-        if gap_freq > max_gap:
-            gap_positions.append(i)
-
-    return gap_positions
-
-
-def calculateSetError(seq_list, ref_seq, ignore_chars=default_missing_chars,
-                      score_dict=getScoreDict(n_score=0, gap_score=0)):
-    """
-    Counts the occurrence of nucleotide mismatches from a reference in a set of sequences
-
-    Arguments:
-    seq_list = a list of SeqRecord objects with aligned sequences
-    ref_seq = a SeqRecord object containing the reference sequence to match against
-    ignore_chars = list of characters to exclude from mismatch counts
-    score_dict = optional dictionary of alignment scores as {(char1, char2): score}
-
-    Returns:
-    a float of the error rate for the set
-    """
-    # Count informative characters in reference sequence
-    ref_bases = sum(1 for b in ref_seq if b not in ignore_chars)
-
-    # Return 0 mismatches for single record case
-    if len(seq_list) <= 1:
-        return 0.0
-
-    # Iterate over seq_list and count mismatches
-    total, score = 0, 0
-    for seq in seq_list:
-        seq_bases = sum(1 for a in seq if a not in ignore_chars)
-        total += min(seq_bases, ref_bases)
-        score += sum([score_dict[(a, b)] for a, b in izip(seq, ref_seq)
-                      if a not in ignore_chars and b not in ignore_chars])
-
-    return 1.0 - float(score) / total
-
-
-def processBCQueue(alive, data_queue, result_queue, cons_func, cons_args={}, 
+def processBCQueue(alive, data_queue, result_queue, cons_func, cons_args={},
                    min_count=default_min_count, primer_field=None, primer_freq=None,
                    max_gap=None, max_error=None, max_diversity=None,
                    copy_fields=None, copy_actions=None, delimiter=default_delimiter):